requirements.txt
schedule
pytz
requests
yfinance
gspread
google-auth
pandas
python-dotenv
